change doubles to floats. Double precision is not actually needed

+++ software/NRG_itho_wifi/main/IthoSystem.cpp
@@ -43,8 +43,8 @@ int32_t *resultPtr2410 = nullptr;
 bool i2c_result_updateweb = false;
 
 bool itho_internal_hum_temp = false;
-double ithoHum = 0;
-double ithoTemp = 0;
+float ithoHum = 0;
+float ithoTemp = 0;
 
 int currentIthoDeviceGroup() { return ithoDeviceGroup; }
 int currentIthoDeviceID() { return ithoDeviceID; }
@@ -298,9 +298,9 @@ void processSettingResult(const uint8_t index, const bool loop)
   root["loop"] = loop;
   if (resultPtr2410 != nullptr && ithoSettingsArray != nullptr)
   {
-    double cur = 0.0;
-    double min = 0.0;
-    double max = 0.0;
+    float cur = 0.0;
+    float min = 0.0;
+    float max = 0.0;
 
     if (decodeQuery2410(resultPtr2410, &ithoSettingsArray[index], &cur, &min, &max))
     {
@@ -521,7 +521,7 @@ void sendRemoteCmd(const uint8_t remoteIndex, const IthoCommand command)
   i2c_header[8] = (curtime >> 8) & 0xFF;
   i2c_header[9] = curtime & 0xFF;
 
-  const int *id = virtualRemotes.getRemoteIDbyIndex(remoteIndex);
+  const uint8_t *id = virtualRemotes.getRemoteIDbyIndex(remoteIndex);
   i2c_header[11] = *id;
   i2c_header[12] = *(id + 1);
   i2c_header[13] = *(id + 2);
@@ -833,7 +833,7 @@ void sendQueryStatus(bool updateweb)
         }
         if (ithoStat.type == ithoDeviceStatus::is_float)
         {
-          double t = ithoStat.value.floatval * ithoStat.divider;
+          float t = ithoStat.value.floatval * ithoStat.divider;
           if (static_cast<uint32_t>(t) == tempVal) // better compare needed of float val, worst case this will result in an extra update of the value, so limited impact
           {
             ithoStat.updated = 0;
@@ -846,7 +846,7 @@ void sendQueryStatus(bool updateweb)
               // interpret raw bytes as signed integer
               tempVal = cast_to_signed_int(tempVal, ithoStat.length);
             }
-            ithoStat.value.floatval = static_cast<double>(tempVal) / ithoStat.divider;
+            ithoStat.value.floatval = static_cast<float>(tempVal) / ithoStat.divider;
           }
         }
 
@@ -1353,7 +1353,7 @@ void sendQuery31D9(bool updateweb)
       }
     }
 
-    double tempVal = i2cbuf[1 + dataStart] / 2.0;
+    float tempVal = i2cbuf[1 + dataStart] / 2.0;
     ithoDeviceMeasurements sTemp = {labels31D9[0], ithoDeviceMeasurements::is_float, {.floatval = tempVal}, 1};
     ithoInternalMeasurements.push_back(sTemp);
 
@@ -1576,9 +1576,9 @@ int32_t *sendQuery2410(uint8_t index, bool updateweb)
       }
       else
       {
-        snprintf(tempbuffer0, sizeof(tempbuffer0), "%.1f", static_cast<double>((int32_t)val0) / ithoSettingsArray[index].divider);
-        snprintf(tempbuffer1, sizeof(tempbuffer1), "%.1f", static_cast<double>((int32_t)val1) / ithoSettingsArray[index].divider);
-        snprintf(tempbuffer2, sizeof(tempbuffer2), "%.1f", static_cast<double>((int32_t)val2) / ithoSettingsArray[index].divider);
+        snprintf(tempbuffer0, sizeof(tempbuffer0), "%.1f", static_cast<float>((int32_t)val0) / ithoSettingsArray[index].divider);
+        snprintf(tempbuffer1, sizeof(tempbuffer1), "%.1f", static_cast<float>((int32_t)val1) / ithoSettingsArray[index].divider);
+        snprintf(tempbuffer2, sizeof(tempbuffer2), "%.1f", static_cast<float>((int32_t)val2) / ithoSettingsArray[index].divider);
       }
       jsonSysmessage("itho2410cur", tempbuffer0);
       jsonSysmessage("itho2410min", tempbuffer1);
@@ -1600,7 +1600,7 @@ int32_t *sendQuery2410(uint8_t index, bool updateweb)
   return values;
 }
 
-bool decodeQuery2410(int32_t *ptr, ithoSettings *setting, double *cur, double *min, double *max)
+bool decodeQuery2410(int32_t *ptr, ithoSettings *setting, float *cur, float *min, float *max)
 {
   if (*(ptr + 0) == 0x5555AAAA && *(ptr + 1) == 0xAAAA5555 && *(ptr + 2) == 0xFFFFFFFF)
   {
@@ -1613,9 +1613,9 @@ bool decodeQuery2410(int32_t *ptr, ithoSettings *setting, double *cur, double *m
   int64_t b = cast_raw_bytes_to_int(ptr + 1, len, is_signed);
   int64_t c = cast_raw_bytes_to_int(ptr + 2, len, is_signed);
 
-  *cur = static_cast<double>(static_cast<int32_t>(a));
-  *min = static_cast<double>(static_cast<int32_t>(b));
-  *max = static_cast<double>(static_cast<int32_t>(c));
+  *cur = static_cast<float>(static_cast<int32_t>(a));
+  *min = static_cast<float>(static_cast<int32_t>(b));
+  *max = static_cast<float>(static_cast<int32_t>(c));
 
   if (setting->type == ithoSettings::is_float)
   {
@@ -1796,7 +1796,7 @@ void filterReset()
   command[8] = (curtime >> 8) & 0xFF;
   command[9] = curtime & 0xFF;
 
-  const int *id = virtualRemotes.getRemoteIDbyIndex(0);
+  const uint8_t *id = virtualRemotes.getRemoteIDbyIndex(0);
   command[11] = *id;
   command[12] = *(id + 1);
   command[13] = *(id + 2);
